llm_non_compliance_checker/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
cleo==1.0.0a5
crashtest==0.3.1
distlib==0.3.6
Django==4.2.7
djangorestframework==3.14.0
dulwich==0.20.50
filelock==3.8.2
frozenlist==1.4.0
html5lib==1.1
idna==3.3
importlib-metadata==4.13.0
importlib-resources==5.10.1
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.11.0
lockfile==0.12.2
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.4
openai==0.28.1
packaging==22.0
pexpect==4.8.0
pkginfo==1.9.2
pkgutil-resolve-name==1.3.10
platformdirs==2.6.0
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyrsistent==0.19.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.26.0
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomlkit==0.11.6
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==1.26.7
virtualenv==20.17.1
webencodings==0.5.1
xattr==0.9.9
yarl==1.9.2
zipp==3.11.0
